extract http endpoints to different file
